separate directory into its own file
Now that we have more file types it doesn't make sense that Directory
lives together with DmaFile.
