+++ scipio/src/io/dma_file.rs
@@ -6,7 +6,6 @@
 use crate::io::read_result::ReadResult;
 use crate::io::scipio_file::ScipioFile;
 use crate::parking::Reactor;
-use crate::sys;
 use crate::sys::sysfs;
 use crate::sys::{DmaBuffer, PollableStatus, SourceType};
 use std::io;
@@ -20,101 +19,6 @@ pub(crate) fn align_up(v: u64, align: u64) -> u64 {
 
 pub(crate) fn align_down(v: u64, align: u64) -> u64 {
     v & !(align - 1)
-}
-
-#[derive(Debug)]
-/// A directory representation where asynchronous operations can be issued
-pub struct Directory {
-    file: ScipioFile,
-}
-
-impl AsRawFd for Directory {
-    fn as_raw_fd(&self) -> RawFd {
-        self.file.as_raw_fd()
-    }
-}
-
-impl Directory {
-    /// Try creating a clone of this Directory.
-    ///
-    /// The new object has a different file descriptor and has to be
-    /// closed separately.
-    pub fn try_clone(&self) -> io::Result<Directory> {
-        let fd = enhanced_try!(
-            sys::duplicate_file(self.file.as_raw_fd()),
-            "Cloning directory",
-            self.file
-        )?;
-        let file = unsafe { ScipioFile::from_raw_fd(fd as _) }.with_path(self.file.path.clone());
-        Ok(Directory { file })
-    }
-
-    /// Synchronously open this directory.
-    pub fn sync_open<P: AsRef<Path>>(path: P) -> io::Result<Directory> {
-        let path = path.as_ref().to_owned();
-        let flags = libc::O_CLOEXEC | libc::O_DIRECTORY;
-        let fd = enhanced_try!(
-            sys::sync_open(&path, flags, 0o755),
-            "Synchronously opening directory",
-            Some(&path),
-            None
-        )?;
-        let file = unsafe { ScipioFile::from_raw_fd(fd as _) }.with_path(Some(path));
-        Ok(Directory { file })
-    }
-
-    /// Asynchronously open the directory at path
-    pub async fn open<P: AsRef<Path>>(path: P) -> io::Result<Directory> {
-        let path = path.as_ref().to_owned();
-        let flags = libc::O_DIRECTORY | libc::O_CLOEXEC;
-        let source = Reactor::get().open_at(-1, &path, flags, 0o755);
-        let fd = enhanced_try!(
-            source.collect_rw().await,
-            "Opening directory",
-            Some(&path),
-            None
-        )?;
-        let file = unsafe { ScipioFile::from_raw_fd(fd as _) }.with_path(Some(path));
-        Ok(Directory { file })
-    }
-
-    /// Similar to create() in the standard library, but returns a DMA file
-    pub fn sync_create<P: AsRef<Path>>(path: P) -> io::Result<Directory> {
-        let path = path.as_ref().to_owned();
-        enhanced_try!(
-            match std::fs::create_dir(&path) {
-                Ok(_) => Ok(()),
-                Err(x) => {
-                    match x.kind() {
-                        std::io::ErrorKind::AlreadyExists => Ok(()),
-                        _ => Err(x),
-                    }
-                }
-            },
-            "Synchronously creating directory",
-            Some(&path),
-            None
-        )?;
-        Self::sync_open(&path)
-    }
-
-    /// Returns an iterator to the contents of this directory
-    pub fn sync_read_dir(&self) -> io::Result<std::fs::ReadDir> {
-        let path = self.file.path_required("read directory")?;
-        enhanced_try!(std::fs::read_dir(path), "Reading a directory", self.file)
-    }
-
-    /// Issues fdatasync into the underlying file.
-    pub async fn sync(&self) -> io::Result<()> {
-        let source = Reactor::get().fdatasync(self.as_raw_fd());
-        source.collect_rw().await?;
-        Ok(())
-    }
-
-    /// Closes this DMA file.
-    pub async fn close(self) -> io::Result<()> {
-        self.file.close().await
-    }
 }
 
 #[derive(Debug)]
